+++ apps/javascript/crdt/text-editors/tiptap/tiptap-crdt-demo/src/components/document/tiptap/index.js
@@ -1,0 +1,13 @@
+/**
+ * TipTap Component Index
+ *
+ * Re-exports for the TipTap editor component.
+ */
+
+export { createTipTapEditor } from './tiptap.js';
+export { initialContent } from './constants.js';
+export { InlineH1, InlineH2, InlineH3 } from './extensions.js';
+export { createBubbleMenuToolbar } from './ui/bubble-menu-toolbar.js';
+export { createToolbarButton } from './ui/toolbar-button.js';
+export { createToolbarDivider } from './ui/toolbar-divider.js';
+
