+++ apps/javascript/crdt/text-editors/tiptap/tiptap-crdt-demo/src/components/document/tiptap/ui/index.js
@@ -1,0 +1,10 @@
+/**
+ * UI Components Index
+ *
+ * Export all UI components for the TipTap editor.
+ */
+
+export { createToolbarButton } from './toolbar-button.js';
+export { createToolbarDivider } from './toolbar-divider.js';
+export { createBubbleMenuToolbar } from './bubble-menu-toolbar.js';
+
