+++ apps/javascript/crdt/text-editors/tiptap/tiptap-crdt-demo/src/components/velt/index.js
@@ -1,0 +1,10 @@
+/**
+ * Velt Components Index
+ *
+ * Export all Velt-related components and functions.
+ */
+
+export { createVeltCollaboration } from './velt-collaboration.js';
+export { initializeVeltDocument, getCurrentDocument, initializeDocument } from './velt-initialize-document.js';
+export { initializeVeltUser, getUser, initializeUser } from './velt-initialize-user.js';
+export { createVeltTools, configureNotificationsTool } from './velt-tools.js';
