+++ apps/javascript/crdt/text-editors/tiptap/tiptap-crdt-demo/styles/globals.css
@@ -57,100 +57,3 @@
     @apply bg-background text-foreground;
   }
 }
-
-/* TipTap Editor Styles - Matching Figma Design */
-.tiptap-editor-content {
-  font-family: 'Geist Mono', monospace;
-  font-size: 12px;
-  font-weight: 400;
-  line-height: 1.5;
-  color: white;
-}
-
-.tiptap-editor-content h1 {
-  font-family: 'Urbanist', sans-serif !important;
-  font-size: 32px !important;
-  font-weight: 700 !important;
-  line-height: 1.5 !important;
-  color: white !important;
-  opacity: 1 !important;
-  background: transparent !important;
-  margin-bottom: 16px;
-}
-
-.tiptap-editor-content h2 {
-  font-family: 'Urbanist', sans-serif !important;
-  font-size: 20px !important;
-  font-weight: 700 !important;
-  line-height: 1.5 !important;
-  color: white !important;
-  opacity: 1 !important;
-  background: transparent !important;
-  margin-bottom: 16px;
-}
-
-.tiptap-editor-content h3 {
-  font-family: 'Urbanist', sans-serif !important;
-  font-size: 16px !important;
-  font-weight: 700 !important;
-  line-height: 1.5 !important;
-  color: white !important;
-  opacity: 1 !important;
-  background: transparent !important;
-  margin-bottom: 16px;
-}
-
-.tiptap-editor-content p {
-  margin-bottom: 16px;
-  opacity: 0.8;
-}
-
-.tiptap-editor-content .ProseMirror {
-  outline: none;
-}
-
-.tiptap-editor-content .ProseMirror:focus {
-  outline: none;
-}
-
-/* Helper Text Animation */
-@keyframes pulse-subtle {
-  0%, 100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.7;
-  }
-}
-
-.animate-pulse-subtle {
-  animation: pulse-subtle 2s cubic-bezier(0.4, 0, 0.6, 1) infinite;
-}
-
-/* Collaboration cursor styling - Dark theme adjustments */
-.collaboration-cursor__caret, .collaboration-carets__caret {
-  border-left: 1px solid #ffffff;
-  border-right: 1px solid #ffffff;
-  margin-left: -1px;
-  margin-right: -1px;
-  pointer-events: none;
-  position: relative;
-  word-break: normal;
-}
-
-/* Render the username above the caret */
-.collaboration-cursor__label, .collaboration-carets__label {
-  border-radius: 3px 3px 3px 0;
-  color: #0d0d0d;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 600;
-  left: -1px;
-  line-height: normal;
-  padding: 0.1rem 0.3rem;
-  position: absolute;
-  top: -1.4em;
-  user-select: none;
-  white-space: nowrap;
-}
-
