+++ apps/javascript/crdt/text-editors/tiptap/tiptap-crdt-demo/styles/tiptap.css
@@ -1,0 +1,144 @@
+/**
+ * TipTap Editor Styles
+ *
+ * Custom styles for the TipTap editor component.
+ */
+
+/* TipTap Editor Styles - Matching Figma Design */
+.tiptap-editor-content {
+  font-family: 'Geist Mono', monospace;
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 1.5;
+  color: white;
+}
+
+.tiptap-editor-content h1 {
+  font-family: 'Urbanist', sans-serif !important;
+  font-size: 32px !important;
+  font-weight: 700 !important;
+  line-height: 1.5 !important;
+  color: white !important;
+  opacity: 1 !important;
+  background: transparent !important;
+  margin-bottom: 16px;
+}
+
+.tiptap-editor-content h2 {
+  font-family: 'Urbanist', sans-serif !important;
+  font-size: 20px !important;
+  font-weight: 700 !important;
+  line-height: 1.5 !important;
+  color: white !important;
+  opacity: 1 !important;
+  background: transparent !important;
+  margin-bottom: 16px;
+}
+
+.tiptap-editor-content h3 {
+  font-family: 'Urbanist', sans-serif !important;
+  font-size: 16px !important;
+  font-weight: 700 !important;
+  line-height: 1.5 !important;
+  color: white !important;
+  opacity: 1 !important;
+  background: transparent !important;
+  margin-bottom: 16px;
+}
+
+.tiptap-editor-content p {
+  margin-bottom: 16px;
+  opacity: 0.8;
+}
+
+.tiptap-editor-content .ProseMirror {
+  outline: none;
+}
+
+.tiptap-editor-content .ProseMirror:focus {
+  outline: none;
+}
+
+/* ProseMirror editor content */
+.ProseMirror {
+  outline: none;
+  min-height: 100%;
+}
+
+.ProseMirror:focus {
+  outline: none;
+}
+
+/* Placeholder text when editor is empty */
+.ProseMirror p.is-editor-empty:first-child::before {
+  color: rgba(255, 255, 255, 0.3);
+  content: attr(data-placeholder);
+  float: left;
+  height: 0;
+  pointer-events: none;
+}
+
+/* Bubble menu styling */
+.bubble-menu-container {
+  z-index: 50;
+}
+
+/* Collaboration cursor styling - Dark theme adjustments */
+.collaboration-cursor__caret,
+.collaboration-carets__caret {
+  border-left: 1px solid #ffffff;
+  border-right: 1px solid #ffffff;
+  margin-left: -1px;
+  margin-right: -1px;
+  pointer-events: none;
+  position: relative;
+  word-break: normal;
+}
+
+/* Render the username above the caret */
+.collaboration-cursor__label,
+.collaboration-carets__label {
+  border-radius: 3px 3px 3px 0;
+  color: #0d0d0d;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 600;
+  left: -1px;
+  line-height: normal;
+  padding: 0.1rem 0.3rem;
+  position: absolute;
+  top: -1.4em;
+  user-select: none;
+  white-space: nowrap;
+}
+
+/* Velt comment highlight styles */
+[data-velt-comment-id] {
+  background-color: rgba(255, 195, 28, 0.2);
+  border-bottom: 2px solid rgba(255, 195, 28, 0.6);
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+}
+
+[data-velt-comment-id]:hover {
+  background-color: rgba(255, 195, 28, 0.3);
+}
+
+[data-velt-comment-id].active {
+  background-color: rgba(255, 195, 28, 0.4);
+}
+
+/* Helper Text Animation */
+@keyframes pulse-subtle {
+  0%,
+  100% {
+    opacity: 1;
+  }
+  50% {
+    opacity: 0.7;
+  }
+}
+
+.animate-pulse-subtle {
+  animation: pulse-subtle 2s cubic-bezier(0.4, 0, 0.6, 1) infinite;
+}
